requirements.txt
Django==5.0.3
djangorestframework==3.15.1
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
djangorestframework-simplejwt==5.5.0
celery